Move miner to minerrepo
